validating meta fields to prevent SQL injection
