+++ integrations/oracle/tests/test_filter_translator.py
@@ -2,6 +2,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+import pytest
+
 from haystack_integrations.document_stores.oracle.filters import FilterTranslator
 
 
@@ -134,3 +136,24 @@ def test_param_counter_increments_correctly():
         }
     )
     assert set(params.keys()) == {"p0", "p1", "p2"}
+
+
+def test_sql_injection_via_field_name_raises():
+    with pytest.raises(ValueError, match="Invalid metadata field name"):
+        _translate({"field": "meta.x') = 'x' OR 1=1 OR ('a", "operator": "==", "value": "y"})
+
+
+def test_sql_injection_via_fallback_field_raises():
+    # Injection via the non-meta fallback path
+    with pytest.raises(ValueError, match="Invalid metadata field name"):
+        _translate({"field": "x') OR 1=1--", "operator": "==", "value": "y"})
+
+
+def test_sql_injection_in_not_in_raises():
+    with pytest.raises(ValueError, match="Invalid metadata field name"):
+        _translate({"field": "meta.x') OR 1=1--", "operator": "in", "value": ["a", "b"]})
+
+
+def test_valid_nested_field_path_accepted():
+    sql, _ = _translate({"field": "meta.author.city", "operator": "==", "value": "Berlin"})
+    assert "'$.author.city'" in sql
